Rename tests.util to tests.utils (like twill.utils)
Better to be consistent with the names.
