+++ tests/test_form.py
@@ -4,7 +4,7 @@ import twill
 from twill import namespaces, commands
 from twill.errors import TwillAssertionError, TwillException
 
-from .util import execute_script
+from .utils import execute_script
 
 
 def test(url):
